Change Read and ReadToCount to return size_t instead of int64_t.

+++ sdk/core/azure-core/inc/azure/core/io/body_stream.hpp
@@ -77,7 +77,7 @@ namespace Azure { namespace Core { namespace IO {
      *
      * @return Number of bytes read.
      */
-    int64_t Read(
+    size_t Read(
         uint8_t* buffer,
         size_t count,
         Azure::Core::Context const& context = Azure::Core::Context())
@@ -96,7 +96,7 @@ namespace Azure { namespace Core { namespace IO {
      *
      * @return Number of bytes read.
      */
-    int64_t ReadToCount(
+    size_t ReadToCount(
         uint8_t* buffer,
         size_t count,
         Azure::Core::Context const& context = Azure::Core::Context());
